requirements.txt
torch>=2.0
python-chess
numpy
pexpect
